Connor dev2

+++ src/main/java/rocks/zipcode/atm/bank/AccountData.java
@@ -9,13 +9,14 @@ public final class AccountData {
     private final String name;
     private final String email;
 
-    private final int balance;
+    private final float balance;
 
-    AccountData(int id, String name, String email, int balance) {
+    AccountData(int id, String name, String email, float balance) {
         this.id = id;
         this.name = name;
         this.email = email;
         this.balance = balance;
+
     }
 
     public int getId() {
@@ -30,9 +31,11 @@ public final class AccountData {
         return email;
     }
 
-    public int getBalance() {
+    public float getBalance() {
         return balance;
     }
+
+
 
     @Override
     public String toString() {
